Remember non-existently populated directory files
The directory file population algorithm has been modified a little bit
to allow remembering of populated files that didn't exist.
The NoDualPopulationFileTest ensures that when a file is accessed, it
won't be re-added to the directory if the underlying file system is
modified.

The class name of the marker SakerDirectory has been change to allow
both use-cases it has.