Set up CI (#8) with test fixes
* Update README.md

* Normalize input nio.Paths for LocalFileProvider

It can cause exceptions on *nix, as x/.. is not handled properly.

* Add message Supplier based assertion functions

Tweak StandardIOFormatterTest to display the output on failures.

Increase test timeout to 2 min.

* Use \r\n as the line separator in StandardIOFormatterTest

If we use System.lineSeparator(), then the test is not deterministic
between platforms.

* Fix StandardIOFormatterTest for \n line separator

* Deterministic get(Type) based [Subscript] fix

The get(Type) method that is called for subscripting is now
deterministic. No longer depends on the order of Class.getMethods().