Fix plugin reload handling in editor model (+NOTE v0.8.11)
Note that this commit was made after the version 0.8.11 was published.
This commit only affects files in the saker.build-ide bundle.

The contents of this commit will be part of the IDE plugin updates, but
not part of the saker.build version 0.8.11. That is, the already
published v0.8.11 will not be altered to include this commit.