Make TaskInvocationConfiguration public API, modify FileProvider API
Deprecate the capability reporting of task factories and made a
preferred method that can be used to construct the configuration object
in a single call.

Slightly modified SakerFileProvider API to return sorted collections
from various methods. This is in order for clients to be able to perform
operations more efficiently when possible. (E.g. sorted iteration)