Require surrounding whitespace for operators in build script
This fixes the issue when multiple command line arguments are specified
in a list without a separating comma. Previously:
[
	-flag1
	-flag2
]
Reported an error that flag2 is not a number due to the second - being
interpreted as a binary subtracting operator.

All operators now require surrounding whitespace so they are not
confused with literals.