Support tracking posix file permissions
The feature is mainly for keeping the execute file permissions. Copying
files should keep the posix file permissions so if a compilation output
file is executable, it should keep its executable permission when copied
into other places via delegate files.

Issues related to this were encountered when implementing iOS support.
The clang output executable has the executable permissions set, but when
copied around into the application directories, they were removed. This
commit attempts to solve this issue by tracking the file permissions.
The build task implemenations need to be modified as well to properly
support this.