Cache environment property values on an execution level
The caching is necessary, so if an environment property value is queried
during an execution, then the same result/exception is returned during
the whole execution. This was necessary, as this commit also introduces
the possibility of recalculating the value of an environment property.

Recalculation can be done if the property calculation exception is
caused by an interruption, so if a property computation fails due to the
execution being interrupted, then this failure doesn't get hard cached,
so the next builds should try the computation again.