Rework SakerTaskResult, mainly replaced by ComposedStructuredTaskResult
The intermediate task resolution in SakerTaskResult is superseeded by
ComposedStructuredTaskResult interface, so it's not necessary to have
that functionality duplicated in it.

Remove the SakerTaskResult get() and getDependencyHandle() functions.

Add StructuredTaskResult.resolveComposition functions that resolve the
composed and object structured task results.

Add SupplierForwardingTaskResultDependencyHandle that wraps a lazily
evaluated TaskResultDependencyHandle.