Don't throw exception if build target parameter is missing and not used
If a build target parameter is not set, and isn't used, then don't throw
an exception, as it's not a serious error.

In the end this means that the absence of the input parameter is
detected at the point of use. If it's not used, then it doesn't cause
the build to fail.